+++ tests/unit/commands/test_deploy.py
@@ -38,6 +38,13 @@ def test_deploy_files_only_smoke_default(
     assert deploy_result.exit_code == 0
 
 
+def test_deploy_assets_only_smoke_default(
+    temp_project: Path, mlflow_file_uploader, mock_dbx_file_upload, mock_api_v2_client
+):
+    deploy_result = invoke_cli_runner(["deploy", "--assets-only"])
+    assert deploy_result.exit_code == 0
+
+
 def test_deploy_multitask_smoke(mlflow_file_uploader, mock_dbx_file_upload, mock_api_v2_client):
     samples_path = get_path_with_relation_to_current_file("../deployment-configs/")
     for file_name in ["03-multitask-job.json", "03-multitask-job.yaml"]:
@@ -144,6 +151,18 @@ def test_deploy_only_chosen_jobs(mlflow_file_uploader, mock_dbx_file_upload, moc
     _chosen = [j["name"] for j in deployment_info.payload.workflows][:2]
     deploy_result = invoke_cli_runner(
         ["deploy", "--environment", "default", "--jobs", ",".join(_chosen), "--write-specs-to-file", result_file],
+    )
+    assert deploy_result.exit_code == 0
+    _content = JsonUtils.read(Path(result_file))
+    assert _chosen == [j["name"] for j in _content["default"]["jobs"]]
+
+
+def test_deploy_only_chosen_workflows(mlflow_file_uploader, mock_dbx_file_upload, mock_api_v2_client):
+    result_file = ".dbx/deployment-result.json"
+    deployment_info = ConfigReader(Path("conf/deployment.yml")).get_environment("default")
+    _chosen = [j["name"] for j in deployment_info.payload.workflows][:2]
+    deploy_result = invoke_cli_runner(
+        ["deploy", "--environment", "default", "--workflows", ",".join(_chosen), "--write-specs-to-file", result_file],
     )
     assert deploy_result.exit_code == 0
     _content = JsonUtils.read(Path(result_file))